requirements.txt
autocorrect==0.4.4
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
datefinder==0.7.0
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-cors-headers==3.1.1
django-elasticsearch-dsl==6.4.1
djangorestframework==3.10.3
elasticsearch==6.4.0
elasticsearch-dsl==6.4.0
et-xmlfile==1.0.1
fasttext==0.9.1
guess-language-spirit==0.5.3
idna==2.8
ipaddress==1.0.23
jdcal==1.4.1
langdetect==1.0.7
lxml==4.4.1
nltk==3.4.5
numpy==1.17.2
oauthlib==3.1.0
opencv-python==4.1.2.30
openpyxl==3.0.0
pandas==0.25.1
polyglot==16.7.4
psycopg2-binary==2.8.3
pybind11==2.4.3
pyenchant==2.0.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
pyudemy==0.1
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.2.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
textblob==0.15.3
urllib3==1.25.3
